requirements.txt
Django==1.11.3
djangorestframework==3.7.3
mysqlclient==1.3.12
pytz==2017.3
